Podcast/requirements.txt
fastapi
uvicorn 
flask
python-dotenv 
gnews
google-generativeai 
langgraph